base class for LLM Client + support deepseek
