+++ cfg/llm_client/deepseek.yaml
@@ -1,7 +1,7 @@
 _target_: utils.llm_client.openai.OpenAIClient
 
 # Using llama api. See the available models at https://docs.llama-api.com/quickstart#available-models
-model: llama3-70
+model: deepseek-chat
 temperature: 1.0  # temperature for chat completion
-api_key: ${oc.env:LLAMA_API_KEY,null}
-base_url: https://api.llama-api.com
+api_key: ${oc.env:DEEPSEEK_API_KEY,null}
+base_url: https://api.deepseek.com
