+++ cfg/llm_client/openai.yaml
@@ -2,5 +2,4 @@ _target_: utils.llm_client.openai.OpenAIClient
 
 model: gpt-3.5-turbo # LLM model
 temperature: 1.0  # temperature for chat completion
-api_timeout: 10.0 # timeout for the API request, in seconds
 api_key: ${oc.env:OPENAI_API_KEY,null}
