+++ cfg/llm_client/zhipuai.yaml
@@ -3,5 +3,4 @@ _target_: utils.llm_client.zhipuai.ZhipuAIClient
 # use LLM API provided by Zhipu AI
 model: GLM-3-Turbo
 temperature: 1.0  # temperature for chat completion
-api_timeout: 10.0 # timeout for the API request, in seconds
 api_key: ${oc.env:ZHIPUAI_API_KEY,null}
