+++ utils/utils.py
@@ -1,15 +1,10 @@
-import subprocess
 import os
-import json
 import logging
-import concurrent.futures
-import time
 import re
 import inspect
 import hydra
-from omegaconf import DictConfig
 
-def init_client(cfg: DictConfig):
+def init_client(cfg):
     global client
     if cfg.get("model", None): # for compatibility
         model: str = cfg.get("model")
@@ -24,7 +19,7 @@ def init_client(cfg: DictConfig):
             from utils.llm_client.openai import OpenAIClient
             # We use llama api here. See the available models at https://docs.llama-api.com/quickstart#available-models
             client = OpenAIClient(
-                model, temperature, 
+                model, temperature,
                 api_key = os.getenv('LLAMA_API_KEY'), 
                 base_url = "https://api.llama-api.com",
             )
@@ -70,76 +65,6 @@ def extract_description(response: str) -> tuple[str, str]:
         if desc_string is not None:
             break
     return desc_string
-
-
-def multi_chat_completion(messages_list: list[list[dict]], n, model, temperature):
-    """
-    An example of messages_list:
-    
-    messages_list = [
-        [
-            {"role": "system", "content": "You are a helpful assistant."},
-            {"role": "user", "content": "Hello!"},
-        ],
-        [
-            {"role": "system", "content": "You are a knowledgeable guide."},
-            {"role": "user", "content": "How are you?"},
-        ],
-        [
-            {"role": "system", "content": "You are a witty comedian."},
-            {"role": "user", "content": "Tell me a joke."},
-        ]
-    ]
-    param: n: number of responses to generate for each message in messages_list
-    """
-    # If messages_list is not a list of list (i.e., only one conversation), convert it to a list of list
-    assert isinstance(messages_list, list), "messages_list should be a list."
-    if not isinstance(messages_list[0], list):
-        messages_list = [messages_list]
-    
-    if len(messages_list) > 1:
-        assert n == 1, "Currently, only n=1 is supported for multi-chat completion."
-    
-    if "gpt" not in model:
-        # Transform messages if n > 1
-        messages_list *= n
-        n = 1
-
-    with concurrent.futures.ThreadPoolExecutor() as executor:
-        args = [(n, messages, model, temperature) for messages in messages_list]
-        choices = executor.map(lambda p: chat_completion(*p), args)
-
-    contents: list[str] = []
-    for choice in choices:
-        for c in choice:
-            contents.append(c.message.content)
-    return contents
-
-
-def chat_completion(n: int, messages: list[dict], model: str, temperature: float) -> list[dict]:
-    """
-    Generate n responses using OpenAI Chat Completions API
-    """
-
-    for attempt in range(1000):
-        try:
-            if "gpt" in model:
-                response_cur = client.chat.completions.create(model=model, messages=messages, temperature=temperature, n=n)
-            else:
-                assert n == 1
-                if "GLM" in model:
-                    response_cur = client.chat.completions.create(model=model, messages=messages, temperature=min(temperature, 1.))
-                else:
-                    response_cur = client.chat.completions.create(model=model, messages=messages, temperature=temperature)
-            break
-        except Exception as e:
-            logging.info(f"Attempt {attempt+1} failed with error: {e}")
-            time.sleep(1)
-    if response_cur is None:
-        logging.info("Code terminated due to too many failed attempts!")
-        exit()
-            
-    return response_cur.choices
 
 
 def extract_code_from_generator(content):
